[#26] Array+Ext 제거
